% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometry_utils.R
\name{dbh2agb_exp}
\alias{dbh2agb_exp}
\title{dbh2agb_exp}
\usage{
dbh2agb_exp(a, b, dbh)
}
\arguments{
\item{a}{a}

\item{b}{b}

\item{dbh}{dbh}
}
\description{
Returns exp allometric dead Biomass
}
\author{
Félicien Meunier
}
