% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometry_utils.R
\name{dbh2bd}
\alias{dbh2bd}
\title{dbh2bd}
\usage{
dbh2bd(b1Bs_small, b2Bs_small, b1Bs_large, b2Bs_large, dbh_crit, dbh)
}
\arguments{
\item{b1Bs_small}{b1Bs_small (ED2)}

\item{b2Bs_small}{b2Bs_small (ED2)}

\item{b1Bs_large}{b1Bs_large (ED2)}

\item{b2Bs_large}{b2Bs_large (ED2)}

\item{dbh_crit}{dbh_crit (ED2)}

\item{dbh}{dbh}
}
\description{
Returns allometric dead Biomass
}
\author{
Félicien Meunier
}
