% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometry_utils.R
\name{dbh2bl}
\alias{dbh2bl}
\title{dbh2bl}
\usage{
dbh2bl(b1Bl_small, b2Bl_small, b1Bl_large, b2Bl_large, dbh_crit, dbh)
}
\arguments{
\item{dbh_crit}{dbh_crit (ED2)}

\item{dbh}{dbh}

\item{b1Bs_small}{b1Bl_small (ED2)}

\item{b2Bs_small}{b2Bl_small (ED2)}

\item{b1Bs_large}{b1Bl_large (ED2)}

\item{b2Bs_large}{b2Bl_large (ED2)}
}
\description{
Returns allometric dead Biomass
}
\author{
Félicien Meunier
}
