% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometry_utils.R
\name{dbh2ca}
\alias{dbh2ca}
\title{dbh2ca}
\usage{
dbh2ca(b1Ca, b2Ca, dbh)
}
\arguments{
\item{b1Ca}{b1Ca (ED2)}

\item{b2Ca}{b2Ca (ED2)}

\item{dbh}{dbh}
}
\description{
Returns allometric Ca
}
\author{
Félicien Meunier
}
