% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometry_utils.R
\name{dbh2h}
\alias{dbh2h}
\title{dbh2h}
\usage{
dbh2h(href, b1Ht, b2Ht, dbh, is.tropical = TRUE)
}
\arguments{
\item{href}{href (ED2)}

\item{b1Ht}{b1Ht (ED2)}

\item{b2Ht}{b2Ht (ED2)}

\item{dbh}{dbh}

\item{is.tropical}{is.tropical}
}
\description{
Returns allometric h
}
\author{
Félicien Meunier
}
