% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ED2_utils.R
\name{get_ED_default_pft}
\alias{get_ED_default_pft}
\title{get_ED_default_pft}
\usage{
get_ED_default_pft(xml, var, pft_target = NULL)
}
\arguments{
\item{xml}{xml (ED2)}

\item{var}{varname (ED2)}

\item{pft_target}{pft num (ED2)}
}
\description{
Returns default parameter
}
\author{
Félicien Meunier
}
