% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_processing.R
\name{patchnumber_from_position}
\alias{patchnumber_from_position}
\title{extremum}
\usage{
patchnumber_from_position(x, y, patch_X, patch_Y)
}
\arguments{
\item{x}{x position}

\item{y}{y position}

\item{patch_X}{patch size in X}

\item{patch_Y}{patch size in Y}
}
\description{
Returns a meaningless (but unique) patch number from position x and y
}
\author{
Félicien Meunier
}
